refactor: add delete single cart item service object
